Made subtotal / total items available globally, also fixed /home route in header
